% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.saehb}
\alias{autoplot.saehb}
\title{Autoplot}
\usage{
\method{autoplot}{saehb}(object, ...)
}
\arguments{
\item{object}{HB model}

\item{...}{other argument}
}
\value{
plot
}
\description{
Autoplot
}
\examples{
\dontrun{
 model1 <- hb_beta(
    y ~ x1 + x2 + x3,
    data = unemployment, seed = 1,
    burn.in = 2000, iter.mcmc = 10000, iter.update = 10,
 )
autoplot(model1)
}

}
