% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corr.R
\name{get_corr}
\alias{get_corr}
\title{Get correlation between y and x variabl}
\usage{
get_corr(y, x, method = "pearson", alpha = NULL, filter_cor = NULL, round = 3)
}
\arguments{
\item{y}{a numeric vector}

\item{x}{a numeric matrix or data.frame}

\item{method}{a character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.}

\item{alpha}{significance level maximum}

\item{filter_cor}{filter minimum correlation}

\item{round}{integer indicating the number of decimal places}
}
\value{
data.frame contains two columns: estimation correlation and p-value of the t-test.
}
\description{
Get correlation between y and x variabl
}
\examples{
get_corr(unemployment$y, x = unemployment[-c(1:2)])
}
