% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hb_beta.R
\name{hb_beta}
\alias{hb_beta}
\title{Small Area Estimation using Hierarchical Bayesian under Beta Distribution}
\usage{
hb_beta(
  formula,
  iter.update = 3,
  iter.mcmc = 10000,
  coef,
  var.coef,
  thin = 3,
  burn.in = 2000,
  tau.u = 1,
  data,
  seed = 1,
  quiet = TRUE,
  plot = TRUE
)
}
\arguments{
\item{formula}{an object of class formula that contains a description of the model to be fitted. The variables included in the formula must be contained in the data.}

\item{iter.update}{Number of updates with default 3}

\item{iter.mcmc}{Number of total iterations per chain with default 10000}

\item{coef}{a vector contains prior initial value of Coefficient of Regression Model for fixed effect with default vector of 0 with the length of the number of regression coefficients}

\item{var.coef}{a vector contains prior initial value of variance of Coefficient of Regression Model with default vector of 1 with the length of the number of regression coefficients}

\item{thin}{Thinning rate, must be a positive integer with default 2}

\item{burn.in}{Number of iterations to discard at the beginning with default 2000}

\item{tau.u}{Prior initial value of inverse of Variance of area random effect with default 1}

\item{data}{data frame containing the variables named in \code{formula}}

\item{seed}{number used to initialize a pseudorandom number generator (default seed = 1). The random number generator method used is "base::Wichmann-Hill".}

\item{quiet}{if TRUE, then messages generated during compilation will be suppressed (default TRUE).}

\item{plot}{if TRUE, the autocorrelation, trace, and density plots will be generated (default TRUE).}
}
\value{
The function returns a list with the following objects : Estimation \code{Est}, random effect variance \code{refVar}, beta coefficient \code{Coefficient} and MCMC result \code{result_mcmc}
}
\description{
Small Area Estimation using Hierarchical Bayesian under Beta Distribution
}
\examples{
\dontrun{
 model1 <- hb_beta(
    y ~ x1 + x2 + x3,
    data = unemployment, seed = 1,
    burn.in = 2000, iter.mcmc = 10000, iter.update = 10,
 )
}
}
